+++ src/main/java/dunn/ted/java/instateam/service/RoleService.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface RoleService {
 
     List<Role> findAll();
-    Role findById(Long id);
+    Role findById(int id);
     void save(Role role);
 
 }
